+++ apps/common/src/main/scala/org/lfdecentralizedtrust/splice/store/AppStoreWithIngestion.scala
@@ -4,7 +4,7 @@
 package org.lfdecentralizedtrust.splice.store
 
 import org.lfdecentralizedtrust.splice.environment.SpliceLedgerConnection
-import org.lfdecentralizedtrust.splice.store.AppStore
+import org.lfdecentralizedtrust.splice.store.AppStoreWithIngestion.SpliceLedgerConnectionPriority
 
 /** A trait for stores that have been wired up with an ingestion pipeline.
   *
@@ -19,6 +19,21 @@ trait AppStoreWithIngestion[Store <: AppStore] {
   /** The store setup with ingestion. */
   def store: Store
 
-  /** A ledger connection whose command submission waits for ingestion into the store. */
-  def connection: SpliceLedgerConnection
+  /** A ledger connection whose command submission waits for ingestion into the store.
+    * @param submissionPriority affects how aggressive is the circuit breaker in stopping submissions.
+    */
+  def connection(submissionPriority: SpliceLedgerConnectionPriority): SpliceLedgerConnection
+}
+
+object AppStoreWithIngestion {
+
+  sealed trait SpliceLedgerConnectionPriority
+
+  object SpliceLedgerConnectionPriority {
+
+    case object High extends SpliceLedgerConnectionPriority
+    case object Medium extends SpliceLedgerConnectionPriority
+    case object Low extends SpliceLedgerConnectionPriority
+
+  }
 }