+++ apps/sv/frontend/src/__tests__/governance/forms/grant-revoke-featured-app-form.test.tsx
@@ -11,6 +11,8 @@ import { Wrapper } from '../../helpers';
 import { dateTimeFormatISO } from '@lfdecentralizedtrust/splice-common-frontend-utils';
 import dayjs from 'dayjs';
 import { GrantRevokeFeaturedAppForm } from '../../../components/forms/GrantRevokeFeaturedAppForm';
+import { server, svUrl } from '../../setup/setup';
+import { rest } from 'msw';
 
 describe('SV user can', () => {
   test('login and see the SV party ID', async () => {
@@ -37,10 +39,7 @@ describe('Grant Featured App Form', () => {
   test('should render all Form components', () => {
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_GrantFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_GrantFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -75,10 +74,7 @@ describe('Grant Featured App Form', () => {
 
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_GrantFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_GrantFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -118,10 +114,7 @@ describe('Grant Featured App Form', () => {
     const user = userEvent.setup();
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_GrantFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_GrantFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -131,14 +124,12 @@ describe('Grant Featured App Form', () => {
     const thePast = dayjs().subtract(1, 'day').format(dateTimeFormatISO);
     const theFuture = dayjs().add(1, 'day').format(dateTimeFormatISO);
 
-    await user.clear(expiryDateInput);
     await user.type(expiryDateInput, thePast);
 
     await waitFor(() => {
       expect(screen.queryByText('Expiration must be in the future')).toBeDefined();
     });
 
-    await user.clear(expiryDateInput);
     await user.type(expiryDateInput, theFuture);
 
     await waitFor(() => {
@@ -151,10 +142,7 @@ describe('Grant Featured App Form', () => {
 
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_GrantFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_GrantFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -164,10 +152,7 @@ describe('Grant Featured App Form', () => {
     const expiryDate = dayjs().add(1, 'week');
     const effectiveDate = expiryDate.subtract(1, 'day');
 
-    await user.clear(expiryDateInput);
     await user.type(expiryDateInput, expiryDate.format(dateTimeFormatISO));
-
-    await user.clear(effectiveDateInput);
     await user.type(effectiveDateInput, effectiveDate.format(dateTimeFormatISO));
 
     await waitFor(() => {
@@ -176,7 +161,6 @@ describe('Grant Featured App Form', () => {
 
     const validEffectiveDate = expiryDate.add(1, 'day').format(dateTimeFormatISO);
 
-    await user.clear(effectiveDateInput);
     await user.type(effectiveDateInput, validEffectiveDate);
 
     await waitFor(() => {
@@ -189,10 +173,7 @@ describe('Grant Featured App Form', () => {
 
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_GrantFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_GrantFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -225,10 +206,7 @@ describe('Revoke Featured App Form', () => {
   test('should render all Form components', () => {
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_RevokeFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_RevokeFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -261,10 +239,7 @@ describe('Revoke Featured App Form', () => {
 
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_RevokeFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_RevokeFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -304,10 +279,7 @@ describe('Revoke Featured App Form', () => {
     const user = userEvent.setup();
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_RevokeFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_RevokeFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -317,14 +289,12 @@ describe('Revoke Featured App Form', () => {
     const thePast = dayjs().subtract(1, 'day').format(dateTimeFormatISO);
     const theFuture = dayjs().add(1, 'day').format(dateTimeFormatISO);
 
-    await user.clear(expiryDateInput);
     await user.type(expiryDateInput, thePast);
 
     await waitFor(() => {
       expect(screen.queryByText('Expiration must be in the future')).toBeDefined();
     });
 
-    await user.clear(expiryDateInput);
     await user.type(expiryDateInput, theFuture);
 
     expect(screen.queryByText('Expiration must be in the future')).toBeNull();
@@ -335,10 +305,7 @@ describe('Revoke Featured App Form', () => {
 
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_RevokeFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_RevokeFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -348,10 +315,7 @@ describe('Revoke Featured App Form', () => {
     const expiryDate = dayjs().add(1, 'week');
     const effectiveDate = expiryDate.subtract(1, 'day');
 
-    await user.clear(expiryDateInput);
     await user.type(expiryDateInput, expiryDate.format(dateTimeFormatISO));
-
-    await user.clear(effectiveDateInput);
     await user.type(effectiveDateInput, effectiveDate.format(dateTimeFormatISO));
 
     await waitFor(() => {
@@ -360,7 +324,6 @@ describe('Revoke Featured App Form', () => {
 
     const validEffectiveDate = expiryDate.add(1, 'day').format(dateTimeFormatISO);
 
-    await user.clear(effectiveDateInput);
     await user.type(effectiveDateInput, validEffectiveDate);
 
     await waitFor(() => {
@@ -373,10 +336,7 @@ describe('Revoke Featured App Form', () => {
 
     render(
       <Wrapper>
-        <GrantRevokeFeaturedAppForm
-          onSubmit={() => Promise.resolve()}
-          selectedAction="SRARC_RevokeFeaturedAppRight"
-        />
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_RevokeFeaturedAppRight" />
       </Wrapper>
     );
 
@@ -402,5 +362,96 @@ describe('Revoke Featured App Form', () => {
     await user.click(submitButton);
 
     expect(screen.getByText('Proposal Summary')).toBeDefined();
+  });
+
+  test('should show error on form if submission fails', async () => {
+    server.use(
+      rest.post(`${svUrl}/v0/admin/sv/voterequest/create`, (_, res, ctx) => {
+        return res(ctx.status(503), ctx.json({ error: 'Service Unavailable' }));
+      })
+    );
+
+    const user = userEvent.setup();
+
+    render(
+      <Wrapper>
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_RevokeFeaturedAppRight" />
+      </Wrapper>
+    );
+
+    const actionInput = screen.getByTestId('revoke-featured-app-action');
+    const submitButton = screen.getByTestId('submit-button');
+
+    const summaryInput = screen.getByTestId('revoke-featured-app-summary');
+    expect(summaryInput).toBeDefined();
+    await user.type(summaryInput, 'Summary of the proposal');
+
+    const urlInput = screen.getByTestId('revoke-featured-app-url');
+    expect(urlInput).toBeDefined();
+    await user.type(urlInput, 'https://example.com');
+
+    const providerInput = screen.getByTestId('revoke-featured-app-idValue');
+    expect(providerInput).toBeDefined();
+    await user.type(providerInput, 'abcde12345');
+
+    await user.click(actionInput); // using this to trigger the onBlur event which triggers the validation
+
+    await waitFor(async () => {
+      expect(submitButton.getAttribute('disabled')).toBeNull();
+    });
+
+    await user.click(submitButton); //review proposal
+    await user.click(submitButton); //submit proposal
+
+    expect(screen.getByTestId('proposal-submission-error')).toBeDefined();
+    expect(screen.getByText(/Submission failed/)).toBeDefined();
+    expect(screen.getByText(/Service Unavailable/)).toBeDefined();
+  });
+
+  test('should redirect to governance page after successful submission', async () => {
+    server.use(
+      rest.post(`${svUrl}/v0/admin/sv/voterequest/create`, (_, res, ctx) => {
+        return res(ctx.json({}));
+      })
+    );
+
+    const user = userEvent.setup();
+
+    render(
+      <Wrapper>
+        <GrantRevokeFeaturedAppForm selectedAction="SRARC_RevokeFeaturedAppRight" />
+      </Wrapper>
+    );
+
+    const actionInput = screen.getByTestId('revoke-featured-app-action');
+    const submitButton = screen.getByTestId('submit-button');
+
+    const summaryInput = screen.getByTestId('revoke-featured-app-summary');
+    expect(summaryInput).toBeDefined();
+    await user.type(summaryInput, 'Summary of the proposal');
+
+    const urlInput = screen.getByTestId('revoke-featured-app-url');
+    expect(urlInput).toBeDefined();
+    await user.type(urlInput, 'https://example.com');
+
+    const providerInput = screen.getByTestId('revoke-featured-app-idValue');
+    expect(providerInput).toBeDefined();
+    await user.type(providerInput, 'abcde12345');
+
+    await user.click(actionInput); // using this to trigger the onBlur event which triggers the validation
+
+    await waitFor(async () => {
+      expect(submitButton.getAttribute('disabled')).toBeNull();
+    });
+
+    await user.click(submitButton); //review proposal
+    await user.click(submitButton); //submit proposal
+
+    waitFor(() => {
+      expect(screen.getByText('Action Required')).toBeDefined();
+      expect(screen.getByText('Inflight Votes')).toBeDefined();
+      expect(screen.getByText('Vote History')).toBeDefined();
+      expect(screen.getByText('Successfully submitted the proposal')).toBeDefined();
+    });
   });
 });