+++ apps/sv/frontend/src/components/forms/UpdateSvRewardWeightForm.tsx
@@ -17,36 +17,38 @@ import {
   validateUrl,
   validateWeight,
 } from './formValidators';
-import { createProposalActions, getInitialExpiration } from '../../utils/governance';
+import {
+  createProposalActions,
+  getInitialExpiration,
+  getSvRewardWeight,
+} from '../../utils/governance';
 import { EffectiveDateField } from '../form-components/EffectiveDateField';
 import { CommonProposalFormData } from '../../utils/types';
 import { ProposalSummary } from '../governance/ProposalSummary';
+import { ProposalSubmissionError } from '../form-components/ProposalSubmissionError';
+import { useProposalMutation } from '../../hooks/useProposalMutation';
 
 interface ExtraFormField {
   sv: string;
   weight: string;
 }
 
-type UpdateSvRewardWeightFormData = CommonProposalFormData & ExtraFormField;
+export type UpdateSvRewardWeightFormData = CommonProposalFormData & ExtraFormField;
 
-interface UpdateSvRewardWeightFormProps {
-  onSubmit: (
-    data: UpdateSvRewardWeightFormData,
-    action: ActionRequiringConfirmation
-  ) => Promise<void>;
-}
-
-export const UpdateSvRewardWeightForm: React.FC<UpdateSvRewardWeightFormProps> = props => {
-  const { onSubmit } = props;
+export const UpdateSvRewardWeightForm: React.FC = _ => {
   const dsoInfosQuery = useDsoInfos();
   const initialExpiration = getInitialExpiration(dsoInfosQuery.data);
   const initialEffectiveDate = dayjs(initialExpiration).add(1, 'day');
   const [showConfirmation, setShowConfirmation] = useState(false);
+  const mutation = useProposalMutation();
 
   const svs = useMemo(
     () => dsoInfosQuery.data?.dsoRules.payload.svs.entriesArray() || [],
     [dsoInfosQuery]
   );
+
+  const svPartyId = dsoInfosQuery.data?.svPartyId || '';
+  const currentWeight = getSvRewardWeight(svs, svPartyId);
 
   const svOptions: { key: string; value: string }[] = useMemo(
     () => svs.map(([partyId, svInfo]) => ({ key: svInfo.name, value: partyId })),
@@ -73,7 +75,7 @@ export const UpdateSvRewardWeightForm: React.FC<UpdateSvRewardWeightFormProps> =
   const form = useAppForm({
     defaultValues,
 
-    onSubmit: ({ value }) => {
+    onSubmit: async ({ value }) => {
       const action: ActionRequiringConfirmation = {
         tag: 'ARC_DsoRules',
         value: {
@@ -90,8 +92,9 @@ export const UpdateSvRewardWeightForm: React.FC<UpdateSvRewardWeightFormProps> =
       if (!showConfirmation) {
         setShowConfirmation(true);
       } else {
-        console.log('submit sv reward weight form data: ', value, 'with action:', action);
-        onSubmit(value, action);
+        await mutation.mutateAsync({ formData: value, action }).catch(e => {
+          console.error(`Failed to submit proposal`, e);
+        });
       }
     },
 
@@ -116,6 +119,7 @@ export const UpdateSvRewardWeightForm: React.FC<UpdateSvRewardWeightFormProps> =
             expiryDate={form.state.values.expiryDate}
             effectiveDate={form.state.values.effectiveDate.effectiveDate}
             formType="sv-reward-weight"
+            currentWeight={currentWeight}
             svRewardWeightMember={form.state.values.sv}
             svRewardWeight={form.state.values.weight}
             onEdit={() => setShowConfirmation(false)}
@@ -212,8 +216,12 @@ export const UpdateSvRewardWeightForm: React.FC<UpdateSvRewardWeightFormProps> =
             </form.AppField>
           </>
         )}
+
         <form.AppForm>
+          <ProposalSubmissionError error={mutation.error} />
+
           <form.FormErrors />
+
           <form.FormControls
             showConfirmation={showConfirmation}
             onEdit={() => setShowConfirmation(false)}