+++ apps/sv/src/main/scala/org/lfdecentralizedtrust/splice/sv/automation/delegatebased/ExpireRewardCouponsTrigger.scala
@@ -3,6 +3,7 @@
 
 package org.lfdecentralizedtrust.splice.sv.automation.delegatebased
 
+import com.digitalasset.canton.topology.PartyId
 import org.lfdecentralizedtrust.splice.automation.{
   PollingParallelTaskExecutionTrigger,
   TaskOutcome,
@@ -21,6 +22,7 @@ import com.digitalasset.canton.tracing.TraceContext
 import com.digitalasset.canton.util.ShowUtil.*
 import io.opentelemetry.api.trace.Tracer
 import org.apache.pekko.stream.Materializer
+import org.lfdecentralizedtrust.splice.store.AppStoreWithIngestion.SpliceLedgerConnectionPriority
 import org.lfdecentralizedtrust.splice.store.PageLimit
 
 import java.util.Optional
@@ -32,6 +34,7 @@ import scala.util.Random
 class ExpireRewardCouponsTrigger(
     override protected val context: TriggerContext,
     override protected val svTaskContext: SvTaskBasedTrigger.Context,
+    ignoredExpiredRewardsPartyIds: Set[PartyId],
 )(implicit
     override val ec: ExecutionContext,
     mat: Materializer,
@@ -48,6 +51,7 @@ class ExpireRewardCouponsTrigger(
       .getExpiredCouponsInBatchesPerRoundAndCouponType(
         dsoRules.domain,
         context.config.enableExpireValidatorFaucet,
+        ignoredExpiredRewardsPartyIds,
         PageLimit.tryCreate(svTaskContext.delegatelessAutomationExpiredRewardCouponBatchSize),
       )
       // We select at most parallelism batches per round as  processing more than that would most likely just hit contention
@@ -57,7 +61,7 @@ class ExpireRewardCouponsTrigger(
 
   override protected def isStaleTask(expiredRewardsTask: ExpiredRewardCouponsBatch)(implicit
       tc: TraceContext
-  ): Future[Boolean] = store.multiDomainAcsStore.hasArchived(
+  ): Future[Boolean] = store.multiDomainAcsStore.containsArchived(
     expiredRewardsTask.validatorCoupons ++ expiredRewardsTask.appCoupons ++ expiredRewardsTask.validatorLivenessActivityRecords ++ expiredRewardsTask.svRewardCoupons
   )
 
@@ -177,7 +181,8 @@ class ExpireRewardCouponsTrigger(
     for {
       _ <- Future.sequence(
         cmds.map(cmd =>
-          svTaskContext.connection
+          svTaskContext
+            .connection(SpliceLedgerConnectionPriority.Low)
             .submit(
               Seq(store.key.svParty),
               Seq(store.key.dsoParty),