+++ apps/sv/src/main/scala/org/lfdecentralizedtrust/splice/sv/store/db/DbSvDsoStore.scala
@@ -442,6 +442,7 @@ class DbSvDsoStore(
   override def listAppRewardCouponsGroupedByRound(
       domain: SynchronizerId,
       totalCouponsLimit: Limit,
+      ignoredParties: Set[PartyId],
   )(implicit
       tc: TraceContext
   ): Future[Seq[RoundBatch[AppRewardCoupon.ContractId]]] =
@@ -449,11 +450,13 @@ class DbSvDsoStore(
       AppRewardCoupon.COMPANION,
       domain,
       totalCouponsLimit,
+      ignoredParties,
     )
 
   override def listValidatorRewardCouponsGroupedByRound(
       domain: SynchronizerId,
       totalCouponsLimit: Limit,
+      ignoredParties: Set[PartyId],
   )(implicit
       tc: TraceContext
   ): Future[Seq[RoundBatch[ValidatorRewardCoupon.ContractId]]] =
@@ -461,11 +464,13 @@ class DbSvDsoStore(
       ValidatorRewardCoupon.COMPANION,
       domain,
       totalCouponsLimit,
+      ignoredParties,
     )
 
   override def listValidatorFaucetCouponsGroupedByRound(
       domain: SynchronizerId,
       totalCouponsLimit: Limit,
+      ignoredParties: Set[PartyId],
   )(implicit
       tc: TraceContext
   ): Future[Seq[RoundBatch[ValidatorFaucetCoupon.ContractId]]] =
@@ -473,11 +478,13 @@ class DbSvDsoStore(
       ValidatorFaucetCoupon.COMPANION,
       domain,
       totalCouponsLimit,
+      ignoredParties,
     )
 
   override def listValidatorLivenessActivityRecordsGroupedByRound(
       domain: SynchronizerId,
       totalCouponsLimit: Limit,
+      ignoredParties: Set[PartyId],
   )(implicit
       tc: TraceContext
   ): Future[Seq[RoundBatch[ValidatorLivenessActivityRecord.ContractId]]] =
@@ -485,33 +492,41 @@ class DbSvDsoStore(
       ValidatorLivenessActivityRecord.COMPANION,
       domain,
       totalCouponsLimit,
+      ignoredParties,
     )
 
   override def listSvRewardCouponsGroupedByRound(
       domain: SynchronizerId,
       totalCouponsLimit: Limit,
+      ignoredParties: Set[PartyId],
   )(implicit tc: TraceContext): Future[Seq[RoundBatch[SvRewardCoupon.ContractId]]] =
     listCouponsGroupedByRound(
       SvRewardCoupon.COMPANION,
       domain,
       totalCouponsLimit,
+      ignoredParties,
     )
 
   private def listCouponsGroupedByRound[C, TCId <: ContractId[_]: ClassTag, T](
       companion: C,
       domain: SynchronizerId,
       totalCouponsLimit: Limit,
+      ignoredParties: Set[PartyId],
   )(implicit
       companionClass: ContractCompanion[C, TCId, T],
       tc: TraceContext,
   ): Future[Seq[SvDsoStore.RoundBatch[TCId]]] = {
     val templateId = companionClass.typeId(companion)
     val opName = s"list${templateId.getEntityName}GroupedByRound"
+    val partyFilter =
+      if (ignoredParties.nonEmpty)
+        (sql"and reward_party not in " ++ inClause(ignoredParties)).toActionBuilder
+      else sql""
     waitUntilAcsIngested {
       for {
         result <- storage
           .query(
-            sql"""
+            (sql"""
                 select reward_round, array_agg(contract_id)
                 from dso_acs_store
                 where store_id = $acsStoreId
@@ -520,18 +535,23 @@ class DbSvDsoStore(
                   and assigned_domain = $domain
                   and reward_party is not null -- otherwise index is not used
                   and reward_round is not null -- otherwise index is not used
+                  """ ++ partyFilter ++ sql"""
                 group by reward_round
                 order by reward_round asc
                 limit ${sqlLimit(totalCouponsLimit)}
-               """.as[(Long, Array[ContractId[ValidatorRewardCoupon]])],
+               """).toActionBuilder.as[(Long, Array[ContractId[ValidatorRewardCoupon]])],
             opName,
           )
-      } yield applyLimit(opName, totalCouponsLimit, result).map { case (round, batch) =>
-        RoundBatch(
-          round,
-          batch.map(cid => companionClass.toContractId(companion, cid.contractId)).toSeq,
-        )
-      }
+      } yield applyLimit(opName, totalCouponsLimit, result)
+        .map { case (round, batch) =>
+          RoundBatch(
+            round,
+            batch
+              .take(totalCouponsLimit.limit)
+              .map(cid => companionClass.toContractId(companion, cid.contractId))
+              .toSeq,
+          )
+        }
     }
   }