+++ cluster/pulumi/common-sv/src/config.ts
@@ -103,7 +103,6 @@ export const SvConfigSchema = z.object({
         .optional(),
       scan: z
         .object({
-          enableImportUpdatesBackfill: z.boolean().optional(),
           rateLimit: z
             .object({
               acs: z
@@ -138,13 +137,3 @@ export const svsConfig = SvConfigSchema.parse(clusterYamlConfig).sv;
 // eslint-disable-next-line
 // @ts-ignore
 export const initialRound = SvConfigSchema.parse(clusterYamlConfig).initialRound;
-
-export const updateHistoryBackfillingValues = svsConfig?.scan?.enableImportUpdatesBackfill
-  ? {
-      updateHistoryBackfilling: {
-        enabled: true,
-        importUpdatesEnabled: true,
-        batchSize: 100,
-      },
-    }
-  : undefined;